+++ src/components/atoms/DisplayErrorMessage.tsx
@@ -1,17 +1,23 @@
 import { Alert } from "antd";
 import { ErrorCode } from "@/@types";
+import { css } from "ss/css";
 
 type Props = {
     error: {
         errorCode: ErrorCode | undefined;
         errorMessage: string;
     };
+    alignLeft?: boolean;
 };
 
-export function DisplayErrorMessage({ error }: Props) {
+export function DisplayErrorMessage({ error, alignLeft = false }: Props) {
     if (!error.errorCode) {
         return;
     }
+
+    const className = css({
+        textAlign: "left"
+    });
 
     return (
         <>
@@ -20,7 +26,7 @@ export function DisplayErrorMessage({ error }: Props) {
                 description={error.errorMessage}
                 type="error"
                 showIcon
-                className="m-5"
+                className={`m-5 ${alignLeft ? className : ""}`}
             />
         </>
     );
